Fix type truncation for larger microseconds (low bpm & very long measures)

+++ src/mainwnd.cpp
@@ -2382,7 +2382,7 @@ mainwnd::signal_action(Glib::IOCondition condition)
 double
 mainwnd::tempo_map_microseconds(unsigned long a_tick)
 {
-    uint32_t hold_microseconds = 0;
+    double hold_microseconds = 0;
 
     list<tempo_mark>::iterator i;
     tempo_mark last_tempo = (*--m_mainperf->m_list_no_stop_markers.end());
@@ -2400,7 +2400,7 @@ mainwnd::tempo_map_microseconds(unsigned long a_tick)
         }
     }
     
-    uint32_t end_tick = a_tick - last_tempo.tick;
+    uint64_t end_tick = a_tick - last_tempo.tick;
    
     return hold_microseconds + ticks_to_delta_time_us (end_tick, last_tempo.bpm, c_ppqn);
 }
